requirements.txt
fastapi
uvicorn[standard]
gradio
python-multipart
opencv-python-headless
numpy
scikit-image
Pillow
ffmpeg-python
